Sneaky reverse proxy handles h2c upgrade
- Copied basically all the h2c upgrade logic from sneaky client into
sneaky reverse proxy
- `httputil.ReverseProxy` attempts to Hijack any requests that result in
a `101 Switching Protocols` from the backend. This is because it is
assuming that the `Upgrade: h2c` header is coming from the original
client, not being injected by the ReverseProxy, like we are doing.
- Since we already have a HTTP/2 request, the Hijack fails (see
https://github.com/golang/go/issues/14797)
- Instead of using a regular `http.Transport`, we wrap it with a
`SneakyTransport`, that will handle the `101` response, thereby
bypassing the doomed-to-fail `ReverseProxy` upgrade logic

Co-authored-by: Greg Cobb <gcobb@vmware.com>